+++ src/server/network/connection/connection.cpp
@@ -113,7 +113,8 @@ void Connection::closeSocket() {
 void Connection::accept(Protocol_ptr protocolPtr) {
 	connectionState = CONNECTION_STATE_IDENTIFYING;
 	protocol = std::move(protocolPtr);
-	g_dispatcher().addEvent([protocol = protocol] { protocol->onConnect(); }, __FUNCTION__, std::chrono::milliseconds(CONNECTION_WRITE_TIMEOUT * 1000).count());
+
+	g_dispatcher().addEvent([eventProtocol = protocol] { eventProtocol->sendLoginChallenge(); }, __FUNCTION__, std::chrono::milliseconds(CONNECTION_WRITE_TIMEOUT * 1000).count());
 
 	acceptInternal(false);
 }
@@ -142,7 +143,7 @@ void Connection::parseProxyIdentification(const std::error_code &error) {
 
 	if (error || connectionState == CONNECTION_STATE_CLOSED) {
 		if (error != asio::error::operation_aborted && error != asio::error::eof && error != asio::error::connection_reset) {
-			g_logger().error("[Connection::parseProxyIdentification] - Read error: {}", error.message());
+			g_logger().debug("[Connection::parseProxyIdentification] - Read error: {}", error.message());
 		}
 		close(FORCE_CLOSE);
 		return;
@@ -217,6 +218,10 @@ void Connection::parseHeader(const std::error_code &error) {
 	}
 
 	uint16_t size = m_msg.getLengthHeader();
+	if (std::dynamic_pointer_cast<ProtocolGame>(protocol)) {
+		size = (size * 8) + 4;
+	}
+
 	if (size == 0 || size > INPUTMESSAGE_MAXSIZE) {
 		close(FORCE_CLOSE);
 		return;
@@ -242,7 +247,7 @@ void Connection::parsePacket(const std::error_code &error) {
 
 	if (error || connectionState == CONNECTION_STATE_CLOSED) {
 		if (error) {
-			g_logger().error("[Connection::parsePacket] - Read error: {}", error.message());
+			g_logger().debug("[Connection::parsePacket] - Read error: {}", error.message());
 		}
 		close(FORCE_CLOSE);
 		return;
@@ -280,7 +285,7 @@ void Connection::parsePacket(const std::error_code &error) {
 			// it doesn't generate any problem because olders protocol don't use 'server sends first' feature
 			m_msg.get<uint32_t>();
 			// Skip protocol ID
-			m_msg.skipBytes(1);
+			m_msg.skipBytes(2);
 		}
 
 		protocol->onRecvFirstMessage(m_msg);