+++ src/server/network/protocol/protocolgame.cpp
@@ -871,7 +871,7 @@ void ProtocolGame::onRecvFirstMessage(NetworkMessage &msg) {
 	g_dispatcher().addEvent([self = getThis(), characterName, accountId, operatingSystem] { self->login(characterName, accountId, operatingSystem); }, __FUNCTION__);
 }
 
-void ProtocolGame::onConnect() {
+void ProtocolGame::sendLoginChallenge() {
 	auto output = OutputMessagePool::getOutputMessage();
 	static std::random_device rd;
 	static std::ranlux24 generator(rd());
@@ -881,7 +881,7 @@ void ProtocolGame::onConnect() {
 	output->skipBytes(sizeof(uint32_t));
 
 	// Packet length & type
-	output->add<uint16_t>(0x0006);
+	output->addByte(0x01);
 	output->addByte(0x1F);
 
 	// Add timestamp & random number
@@ -890,6 +890,7 @@ void ProtocolGame::onConnect() {
 
 	challengeRandom = randNumber(generator);
 	output->addByte(challengeRandom);
+	output->addByte(0x71);
 
 	// Go back and write checksum
 	output->skipBytes(-12);
@@ -1620,12 +1621,11 @@ void ProtocolGame::parseOpenPrivateChannel(NetworkMessage &msg) {
 
 void ProtocolGame::parseAutoWalk(NetworkMessage &msg) {
 	uint8_t numdirs = msg.getByte();
-	if (numdirs == 0 || (msg.getBufferPosition() + numdirs) != (msg.getLength() + 8)) {
+	if (numdirs == 0 || (msg.getBufferPosition() + numdirs) != (msg.getLength() + 6)) {
 		return;
 	}
 
-	std::vector<Direction> path;
-	path.resize(numdirs, DIRECTION_NORTH);
+	std::vector<Direction> path(numdirs, DIRECTION_NORTH);
 	for (size_t i = numdirs; --i < numdirs;) {
 		const uint8_t rawdir = msg.getByte();
 		switch (rawdir) {
@@ -4663,8 +4663,8 @@ void ProtocolGame::sendIcons(const std::unordered_set<PlayerIcon> &iconSet, cons
 		// Send as uint16_t in old protocol
 		msg.add<uint16_t>(static_cast<uint16_t>(icons));
 	} else {
-		// Send as uint32_t in new protocol
-		msg.add<uint32_t>(icons);
+		// Send as uint64_t in new protocol
+		msg.add<uint64_t>(icons);
 		msg.addByte(enumToValue(iconBakragore)); // Icons Bakragore
 	}
 
@@ -4674,7 +4674,7 @@ void ProtocolGame::sendIcons(const std::unordered_set<PlayerIcon> &iconSet, cons
 void ProtocolGame::sendIconBakragore(const IconBakragore icon) {
 	NetworkMessage msg;
 	msg.addByte(0xA2);
-	msg.add<uint32_t>(0); // Send empty normal icons
+	msg.add<uint64_t>(0); // Send empty normal icons
 	msg.addByte(enumToValue(icon));
 	writeToOutputBuffer(msg);
 }
@@ -8597,7 +8597,7 @@ void ProtocolGame::sendOpenStash() {
 		msg.add<uint16_t>(item.first);
 		msg.add<uint32_t>(item.second);
 	}
-	msg.add<uint16_t>(static_cast<uint16_t>(g_configManager().getNumber(STASH_ITEMS) - getStashSize(list)));
+
 	writeToOutputBuffer(msg);
 }
 
@@ -8616,9 +8616,9 @@ void ProtocolGame::parseStashWithdraw(NetworkMessage &msg) {
 		return;
 	}
 
-	auto action = static_cast<Supply_Stash_Actions_t>(msg.getByte());
+	auto action = static_cast<Stash_Actions_t>(msg.getByte());
 	switch (action) {
-		case SUPPLY_STASH_ACTION_STOW_ITEM: {
+		case STASH_ACTION_STOW_ITEM: {
 			Position pos = msg.getPosition();
 			auto itemId = msg.get<uint16_t>();
 			uint8_t stackpos = msg.getByte();
@@ -8626,21 +8626,21 @@ void ProtocolGame::parseStashWithdraw(NetworkMessage &msg) {
 			g_game().playerStowItem(player->getID(), pos, itemId, stackpos, count, false);
 			break;
 		}
-		case SUPPLY_STASH_ACTION_STOW_CONTAINER: {
+		case STASH_ACTION_STOW_CONTAINER: {
 			Position pos = msg.getPosition();
 			auto itemId = msg.get<uint16_t>();
 			uint8_t stackpos = msg.getByte();
 			g_game().playerStowItem(player->getID(), pos, itemId, stackpos, 0, false);
 			break;
 		}
-		case SUPPLY_STASH_ACTION_STOW_STACK: {
+		case STASH_ACTION_STOW_STACK: {
 			Position pos = msg.getPosition();
 			auto itemId = msg.get<uint16_t>();
 			uint8_t stackpos = msg.getByte();
 			g_game().playerStowItem(player->getID(), pos, itemId, stackpos, 0, true);
 			break;
 		}
-		case SUPPLY_STASH_ACTION_WITHDRAW: {
+		case STASH_ACTION_WITHDRAW: {
 			auto itemId = msg.get<uint16_t>();
 			auto count = msg.get<uint32_t>();
 			uint8_t stackpos = msg.getByte();